+++ packages/composables/index.ts
@@ -2,3 +2,4 @@ export * from "./config";
 export * from "./props";
 export * from "./dict";
 export * from "./selected";
+export * from "./option";