+++ packages/constants/props.ts
@@ -1,16 +1,17 @@
 import type { PropType } from "vue-demi";
-import type { Config, DictValue, DictData, DictComponentType } from "../types";
+import type { Config, DictValue, DictData, DictComponentType, DictItem } from "../types";
 
 export const basicProps = {
   value: { type: [String, Number, Array] as PropType<DictValue>, default: "" }, // Vue2
   modelValue: { type: [String, Number, Array] as PropType<DictValue>, default: "" }, // Vue3
+  data: { type: Array as PropType<DictItem[]>, default: () => [] }
+};
+
+export const dictRestProps = {
+  type: { type: String as PropType<DictComponentType>, default: "select" },
   data: { type: [Array, Function] as PropType<DictData>, default: () => [] },
   option: { type: Object as PropType<Config> },
   cache: { type: String, default: "" }
-};
-
-export const typeProps = {
-  type: { type: String as PropType<DictComponentType>, default: "select" }
 };
 
 export const buttonProps = {
@@ -19,6 +20,6 @@ export const buttonProps = {
 
 export const fullProps = {
   ...basicProps,
-  ...typeProps,
+  ...dictRestProps,
   ...buttonProps
 };