+++ CHANGELOG.md
@@ -95,7 +95,7 @@
 
 ### Features
 
-* v-dict,select,radio,check,text ([1c7bed7](https://github.com/SoulLyoko/vue-dictionary/commit/1c7bed70ba3670e7ef247b4c08b98371b219f65f))
+* dict,select,radio,check,text ([1c7bed7](https://github.com/SoulLyoko/vue-dictionary/commit/1c7bed70ba3670e7ef247b4c08b98371b219f65f))